requirements.txt
customtkinter>=5.2.0
requests>=2.28.0
Pillow>=9.0.0
packaging>=23.0
